requirements.txt
pywhatkit>=5.4
twilio>=8.0.0
requests>=2.31.0
python-dotenv>=1.0.0
